product valid added

+++ src/main/java/com/redifoglu/ecommerce/dto/CartDTO.java
@@ -1,11 +1,12 @@
 package com.redifoglu.ecommerce.dto;
 
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CartDTO(Long id,
-                      Double itemTotal,
-                      Double grandTotal,
+                      BigDecimal itemTotal,
+                      BigDecimal grandTotal,
                       CustomerDTO customer,
                       List<ProductDTO> products) {
 }
